Kumar\CKYC-177\Test and Develop MSISDN History API Getting recent MSISDN transaction

+++ src/main/java/com/app/kyc/entity/MsisdnTracking.java
@@ -35,7 +35,7 @@ public class MsisdnTracking {
     private String status;
 
     @Column(name = "service_provider_id ", length = 100)
-    private Integer serviceProviderId ;
+    private Long serviceProviderId ;
 
     @Column(name = "identification_number ", length = 100)
     private String identificationNumber;
